ok3

+++ zbuild/travis/install.sh
@@ -74,11 +74,25 @@ make_conda() {
 
 
     #### Test env install
-    # conda create -n testenv --yes $TO_INSTALL --file zbuild/py36.txt
-    source activate testenv
-    pip install arrow==0.10.0 attrdict==2.0.0 backports.shutil-get-terminal-size==1.0.0 configmy==0.14.87 github3.py==1.2.0 jwcrypto==0.6.0 kmodes==0.9 rope-py3k==0.9.4.post1 tables==3.3.0 tabulate==0.8.2 uritemplate==3.0.0
-    pip install pytest==4.3.0
-    pip install toml
+    if test -e $HOME/miniconda/bin; then
+      export PATH=$HOME/miniconda/bin:$PATH
+      conda update --yes conda
+      source activate testenv
+
+    else
+
+      chmod +x $HOME/download/miniconda.sh
+      $HOME/download/miniconda.sh -b -p $HOME/miniconda;
+      export PATH=$HOME/miniconda/bin:$PATH
+      conda update --yes conda
+
+      conda create -n testenv --yes $TO_INSTALL --file zbuild/py36.txt
+      source activate testenv
+      pip install arrow==0.10.0 attrdict==2.0.0 backports.shutil-get-terminal-size==1.0.0 configmy==0.14.87 github3.py==1.2.0 jwcrypto==0.6.0 kmodes==0.9 rope-py3k==0.9.4.post1 tables==3.3.0 tabulate==0.8.2 uritemplate==3.0.0
+      pip install pytest==4.3.0
+      pip install toml
+    fi;
+
 
     python --version
     python -c "import numpy; print('numpy %s' % numpy.__version__)"
